프로토콜 수정. float -> int

+++ Assets/Scripts/Network/RWGameProtocol/Msg/MsgHitDamage.cs
@@ -7,8 +7,8 @@ namespace RWGameProtocol.Msg
     [StructLayout(LayoutKind.Sequential, Pack = 1)]
     public class MsgHitDamageReq : Serializer<MsgHitDamageReq>
     {
-        // 데미지량
-        public float Damage;
+        // 데미지량. (실제 데미지량은 0.01을 곱해서 사용해야 함)
+        public int Damage;
     }
 
 
@@ -18,8 +18,8 @@ namespace RWGameProtocol.Msg
     {
         public short ErrorCode;
 
-        // 데미지량
-        public float Damage;
+        // 데미지량 (실제 데미지량은 0.01을 곱해서 사용해야 함)
+        public int Damage;
     }
 
     [Serializable]
@@ -29,7 +29,7 @@ namespace RWGameProtocol.Msg
         // 플레이어 UID
         public int PlayerUId;
 
-        // 데미지량
-        public float Damage;
+        // 데미지량 (실제 데미지량은 0.01을 곱해서 사용해야 함)
+        public int Damage;
     }
 }
